Consistency of parameter names, drastic reduction in replicated code, and beginning to massage the DiskIndex and build_disk_index functions to follow suit of the memory indices
