+++ python/src/__init__.py
@@ -1,13 +1,14 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT license.
 
-from ._wrapper import (
-    DiskIndex,
-    DynamicMemoryIndex,
-    StaticMemoryIndex,
-    VectorDType,
+from ._builder import (
     build_disk_index_from_vector_file,
     build_disk_index_from_vectors,
     build_memory_index_from_vector_file,
     numpy_to_diskann_file,
 )
+from ._common import VectorDType
+from ._disk_index import DiskIndex
+from ._diskannpy import INNER_PRODUCT, L2, Metric, defaults
+from ._dynamic_memory_index import DynamicMemoryIndex
+from ._static_memory_index import StaticMemoryIndex
