+++ python/src/_common.py
@@ -1,0 +1,58 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT license.
+
+from typing import Type, TypeVar
+
+import numpy as np
+
+from . import _diskannpy as _native_dap
+
+__ALL__ = ["VectorDType"]
+
+_VALID_DTYPES = [np.single, np.byte, np.ubyte]
+
+VectorDType = TypeVar("VectorDType", Type[np.single], Type[np.ubyte], Type[np.byte])
+
+
+def _get_valid_metric(metric: str) -> _native_dap.Metric:
+    if not isinstance(metric, str):
+        raise ValueError("metric must be a string")
+    if metric.lower() == "l2":
+        return _native_dap.L2
+    elif metric.lower() == "mips":
+        return _native_dap.INNER_PRODUCT
+    else:
+        raise ValueError("metric must be one of 'l2' or 'mips'")
+
+
+def _assert_dtype(vectors, name: str):
+    _assert(
+        vectors in _VALID_DTYPES,
+        f"{name} must be of one of type np.single, np.byte, or np.ubyte",
+    )
+
+
+def _assert_2d(vectors: np.ndarray, name: str):
+    _assert(len(vectors.shape) == 2, f"{name} must be 2d numpy array")
+
+
+__MAX_UINT_VAL = 4_294_967_295
+
+
+def _assert_is_positive_uint32(test_value: int, parameter: str):
+    _assert(
+        0 < test_value < __MAX_UINT_VAL,
+        f"{parameter} must be a positive integer in the uint32 range",
+    )
+
+
+def _assert_is_nonnegative_uint32(test_value: int, parameter: str):
+    _assert(
+        -1 < test_value < __MAX_UINT_VAL,
+        f"{parameter} must be a non-negative integer in the uint32 range",
+    )
+
+
+def _assert(statement_eval: bool, message: str):
+    if not statement_eval:
+        raise ValueError(message)
