+++ python/src/diskann_bindings.cpp
@@ -23,7 +23,7 @@
 #include "pq_flash_index.h"
 #include "utils.h"
 
-PYBIND11_MAKE_OPAQUE(std::vector<unsigned>);
+PYBIND11_MAKE_OPAQUE(std::vector<uint32_t>);
 PYBIND11_MAKE_OPAQUE(std::vector<float>);
 PYBIND11_MAKE_OPAQUE(std::vector<int8_t>);
 PYBIND11_MAKE_OPAQUE(std::vector<uint8_t>);
@@ -37,18 +37,18 @@ typedef WindowsAlignedFileReader PlatformSpecificAlignedFileReader;
 typedef LinuxAlignedFileReader PlatformSpecificAlignedFileReader;
 #endif
 
-template <class T> struct DiskANNIndex
+template <class T> struct DiskIndex
 {
     PQFlashIndex<T> *_pq_flash_index;
     std::shared_ptr<AlignedFileReader> reader;
 
-    DiskANNIndex(diskann::Metric metric)
+    DiskIndex(diskann::Metric metric)
     {
         reader = std::make_shared<PlatformSpecificAlignedFileReader>();
         _pq_flash_index = new PQFlashIndex<T>(reader, metric);
     }
 
-    ~DiskANNIndex()
+    ~DiskIndex()
     {
         delete _pq_flash_index;
     }
@@ -103,16 +103,16 @@ template <class T> struct DiskANNIndex
     }
 
     auto search(py::array_t<T, py::array::c_style | py::array::forcecast> &query, const uint64_t knn,
-                const uint64_t l_search, const uint64_t beam_width)
+                const uint64_t complexity, const uint64_t beam_width)
     {
-        py::array_t<unsigned> ids(knn);
+        py::array_t<uint32_t> ids(knn);
         py::array_t<float> dists(knn);
 
-        std::vector<unsigned> u32_ids(knn);
+        std::vector<uint32_t> u32_ids(knn);
         std::vector<uint64_t> u64_ids(knn);
         QueryStats stats;
 
-        _pq_flash_index->cached_beam_search(query.data(), knn, l_search, u64_ids.data(), dists.mutable_data(),
+        _pq_flash_index->cached_beam_search(query.data(), knn, complexity, u64_ids.data(), dists.mutable_data(),
                                             beam_width, false, &stats);
 
         auto r = ids.mutable_unchecked<1>();
@@ -123,9 +123,9 @@ template <class T> struct DiskANNIndex
     }
 
     auto batch_search(py::array_t<T, py::array::c_style | py::array::forcecast> &queries, const uint64_t num_queries,
-                      const uint64_t knn, const uint64_t l_search, const uint64_t beam_width, const int num_threads)
+                      const uint64_t knn, const uint64_t complexity, const uint64_t beam_width, const int num_threads)
     {
-        py::array_t<unsigned> ids({num_queries, knn});
+        py::array_t<uint32_t> ids({num_queries, knn});
         py::array_t<float> dists({num_queries, knn});
 
         omp_set_num_threads(num_threads);
@@ -135,14 +135,14 @@ template <class T> struct DiskANNIndex
 #pragma omp parallel for schedule(dynamic, 1)
         for (int64_t i = 0; i < (int64_t)num_queries; i++)
         {
-            _pq_flash_index->cached_beam_search(queries.data(i), knn, l_search, u64_ids.data() + i * knn,
+            _pq_flash_index->cached_beam_search(queries.data(i), knn, complexity, u64_ids.data() + i * knn,
                                                 dists.mutable_data(i), beam_width);
         }
 
         auto r = ids.mutable_unchecked();
         for (uint64_t i = 0; i < num_queries; ++i)
             for (uint64_t j = 0; j < knn; ++j)
-                r(i, j) = (unsigned)u64_ids[i * knn + j];
+                r(i, j) = (uint32_t)u64_ids[i * knn + j];
 
         return std::make_pair(ids, dists);
     }
@@ -154,33 +154,43 @@ typedef uint32_t filterT;
 template <class T> struct DynamicInMemIndex
 {
     Index<T, IdT, filterT> *_index;
-    IndexWriteParameters write_params;
+    IndexWriteParameters _write_params;
 
-    DynamicInMemIndex(Metric m, const size_t dim, const size_t max_points, const uint32_t l_build,
-                      const uint32_t build_max_degree, const bool saturate_graph, const uint32_t max_occlusion_size,
-                      const float alpha, const uint32_t num_threads, const uint32_t filter_list_size,
-                      const uint32_t num_frozen_points, const uint32_t initial_search_list_size,
-                      const uint32_t search_threads, const bool concurrent_consolidate)
-        : write_params(IndexWriteParametersBuilder(l_build, build_max_degree)
-                           .with_saturate_graph(saturate_graph)
-                           .with_max_occlusion_size(max_occlusion_size)
-                           .with_alpha(alpha)
-                           .with_num_threads(num_threads)
-                           .with_filter_list_size(filter_list_size)
-                           .with_num_frozen_points(num_frozen_points)
-                           .build())
+    DynamicInMemIndex(Metric m, const size_t dim, const size_t max_points, const uint32_t complexity,
+                      const uint32_t graph_degree, const bool saturate_graph, const uint32_t max_occlusion_size,
+                      const float alpha, const uint32_t num_threads, const uint32_t filter_complexity,
+                      const uint32_t num_frozen_points, const uint32_t initial_search_complexity,
+                      const uint32_t initial_search_threads, const bool concurrent_consolidation,
+                      const std::string &index_path = "")
+        : _write_params(IndexWriteParametersBuilder(complexity, graph_degree)
+                            .with_saturate_graph(saturate_graph)
+                            .with_max_occlusion_size(max_occlusion_size)
+                            .with_alpha(alpha)
+                            .with_num_threads(num_threads)
+                            .with_filter_list_size(filter_complexity)
+                            .with_num_frozen_points(num_frozen_points)
+                            .build())
     {
+        const uint32_t _initial_search_complexity =
+            initial_search_complexity != 0 ? initial_search_complexity : complexity;
+        const uint32_t _initial_search_threads =
+            initial_search_threads != 0 ? initial_search_threads : omp_get_num_threads();
+
         _index = new Index<T>(m, dim, max_points,
-                              true,                     // dynamic_index
-                              write_params,             // used for insert
-                              initial_search_list_size, // used to prepare the scratch space for searching. can / may be
-                                                        // expanded if the search asks for a larger L.
-                              search_threads,           // also used for the scratch space
-                              true,                     // enable_tags
-                              concurrent_consolidate,
+                              true,                       // dynamic_index
+                              _write_params,              // used for insert
+                              _initial_search_complexity, // used to prepare the scratch space for searching. can / may
+                                                          // be expanded if the search asks for a larger L.
+                              _initial_search_threads,    // also used for the scratch space
+                              true,                       // enable_tags
+                              concurrent_consolidation,
                               false,  // pq_dist_build
                               0,      // num_pq_chunks
                               false); // use_opq = false
+        if (index_path != "")
+        {
+            _index->load(index_path.c_str(), _write_params.num_threads, complexity);
+        }
     }
 
     ~DynamicInMemIndex()
@@ -199,17 +209,17 @@ template <class T> struct DynamicInMemIndex
     }
 
     auto search(py::array_t<T, py::array::c_style | py::array::forcecast> &query, const uint64_t knn,
-                const uint64_t l_search)
+                const uint64_t complexity)
     {
         py::array_t<IdT> ids(knn);
         py::array_t<float> dists(knn);
         std::vector<T *> empty_vector;
-        _index->search_with_tags(query.data(), knn, l_search, ids.mutable_data(), dists.mutable_data(), empty_vector);
+        _index->search_with_tags(query.data(), knn, complexity, ids.mutable_data(), dists.mutable_data(), empty_vector);
         return std::make_pair(ids, dists);
     }
 
     auto batch_search(py::array_t<T, py::array::c_style | py::array::forcecast> &queries, const uint64_t num_queries,
-                      const uint64_t knn, const uint64_t l_search, const int num_threads)
+                      const uint64_t knn, const uint64_t complexity, const int num_threads)
     {
         py::array_t<unsigned> ids({num_queries, knn});
         py::array_t<float> dists({num_queries, knn});
@@ -220,7 +230,7 @@ template <class T> struct DynamicInMemIndex
 #pragma omp parallel for schedule(dynamic, 1)
         for (int64_t i = 0; i < (int64_t)num_queries; i++)
         {
-            _index->search_with_tags(queries.data(i), knn, l_search, ids.mutable_data(i), dists.mutable_data(i),
+            _index->search_with_tags(queries.data(i), knn, complexity, ids.mutable_data(i), dists.mutable_data(i),
                                      empty_vector);
         }
 
@@ -229,18 +239,23 @@ template <class T> struct DynamicInMemIndex
 
     auto consolidate_delete()
     {
-        return _index->consolidate_deletes(write_params);
+        return _index->consolidate_deletes(_write_params);
     }
 };
 
 template <class T> struct StaticInMemIndex
 {
     Index<T, IdT, filterT> *_index;
-    const IndexWriteParameters write_params;
 
-    StaticInMemIndex(Metric m, const std::string &data_path, const uint32_t l_build, const uint32_t graph_degree)
-        : write_params(diskann::IndexWriteParametersBuilder(l_build, graph_degree).build())
+    StaticInMemIndex(const Metric m, const std::string &data_path, const std::string &index_prefix,
+                     const uint32_t num_threads, const uint32_t initial_search_complexity)
     {
+        const uint32_t _num_threads = num_threads != 0 ? num_threads : omp_get_num_threads();
+        if (initial_search_complexity == 0)
+        {
+            throw std::runtime_error("initial_search_complexity must be a positive uint32_t");
+        }
+
         size_t ndims, npoints;
         diskann::get_bin_metadata(data_path, npoints, ndims);
         _index = new Index<T>(m, ndims, npoints,
@@ -251,7 +266,7 @@ template <class T> struct StaticInMemIndex
                               0,     // num_pq_chunks
                               false, // use_opq = false
                               0);    // num_frozen_pts = 0
-        _index->build(data_path.c_str(), npoints, write_params);
+        _index->load(index_prefix.c_str(), _num_threads, initial_search_complexity);
     }
 
     ~StaticInMemIndex()
@@ -260,33 +275,48 @@ template <class T> struct StaticInMemIndex
     }
 
     auto search(py::array_t<T, py::array::c_style | py::array::forcecast> &query, const uint64_t knn,
-                const uint64_t l_search)
+                const uint64_t complexity)
     {
         py::array_t<IdT> ids(knn);
         py::array_t<float> dists(knn);
         std::vector<T *> empty_vector;
-        _index->search(query.data(), knn, l_search, ids.mutable_data(), dists.mutable_data());
+        _index->search(query.data(), knn, complexity, ids.mutable_data(), dists.mutable_data());
         return std::make_pair(ids, dists);
     }
 
     auto batch_search(py::array_t<T, py::array::c_style | py::array::forcecast> &queries, const uint64_t num_queries,
-                      const uint64_t knn, const uint64_t l_search, const int num_threads)
+                      const uint64_t knn, const uint64_t complexity, const int num_threads)
     {
+        const uint32_t _num_threads = num_threads != 0 ? num_threads : omp_get_num_threads();
         py::array_t<unsigned> ids({num_queries, knn});
         py::array_t<float> dists({num_queries, knn});
         std::vector<T *> empty_vector;
 
-        omp_set_num_threads(num_threads);
+        omp_set_num_threads(_num_threads);
 
 #pragma omp parallel for schedule(dynamic, 1)
         for (int64_t i = 0; i < (int64_t)num_queries; i++)
         {
-            _index->search(queries.data(i), knn, l_search, ids.mutable_data(i), dists.mutable_data(i));
+            _index->search(queries.data(i), knn, complexity, ids.mutable_data(i), dists.mutable_data(i));
         }
 
         return std::make_pair(ids, dists);
     }
 };
+
+template <typename T>
+void build_disk_index2(const diskann::Metric metric, const std::string &data_file_path,
+                      const std::string &index_prefix_path, const uint32_t complexity, const uint32_t graph_degree,
+                      const double final_index_ram_limit, const double indexing_ram_budget, const uint32_t num_threads,
+                      const uint32_t pq_disk_bytes)
+{
+    std::string params = std::to_string(graph_degree) + " " + std::to_string(complexity) + " " +
+                         std::to_string(final_index_ram_limit) + " " + std::to_string(indexing_ram_budget) + " " +
+                         std::to_string(num_threads);
+    if (pq_disk_bytes > 0)
+        params = params + " " + std::to_string(pq_disk_bytes);
+    diskann::build_disk_index<T>(data_file_path.c_str(), index_prefix_path.c_str(), params.c_str(), metric);
+}
 
 template <typename T, typename TagT = IdT, typename LabelT = filterT>
 void build_in_memory_index(const diskann::Metric &metric, const std::string &vector_bin_path,
@@ -324,6 +354,75 @@ void build_in_memory_index(const diskann::Metric &metric, const std::string &vec
     index.save(index_output_path.c_str());
 }
 
+template <typename T> inline void add_variant(py::module_ &m, const std::string &build_name,
+                                const std::string &class_name) {
+    const std::string build_disk_name = "build_disk_" + build_name + "_index";
+    m.def(build_disk_name.c_str(), &build_disk_index2<T>,
+          py::arg("metric"), py::arg("data_file_path"), py::arg("index_prefix_path"), py::arg("complexity"),
+          py::arg("graph_degree"), py::arg("final_index_ram_limit"), py::arg("indexing_ram_budget"),
+          py::arg("num_threads"), py::arg("pq_disk_bytes"));
+
+    const std::string build_in_memory_name = "build_in_memory_" + build_name + "_index";
+    m.def(build_in_memory_name.c_str(), &build_in_memory_index<float>, py::arg("metric"), py::arg("data_file_path"),
+          py::arg("index_output_path"), py::arg("graph_degree"), py::arg("complexity"), py::arg("alpha"),
+          py::arg("num_threads"), py::arg("use_pq_build"), py::arg("num_pq_bytes"), py::arg("use_opq"),
+          py::arg("label_file") = "", py::arg("universal_label") = "", py::arg("filter_complexity") = 0);
+
+        const std::string static_index = "StaticMemory" + class_name + "Index";
+        py::class_<StaticInMemIndex<T>>(m, static_index.c_str())
+            .def(py::init([](const diskann::Metric metric, const std::string &data_path, const std::string &index_path,
+                             const uint32_t num_threads, const uint32_t initial_search_complexity) {
+                     return std::unique_ptr<StaticInMemIndex<T>>(new StaticInMemIndex<T>(
+                         metric, data_path, index_path, num_threads, initial_search_complexity));
+                 }),
+                 py::arg("metric"), py::arg("data_path"), py::arg("index_path"), py::arg("num_threads"),
+                 py::arg("initial_search_complexity"))
+            .def("search", &StaticInMemIndex<T>::search, py::arg("query"), py::arg("knn"), py::arg("complexity"))
+            .def("batch_search", &StaticInMemIndex<T>::batch_search, py::arg("queries"), py::arg("num_queries"),
+                 py::arg("knn"), py::arg("complexity"), py::arg("num_threads"));
+
+        const std::string dynamic_index = "DynamicMemory" + class_name + "Index";
+        py::class_<DynamicInMemIndex<T>>(m, dynamic_index.c_str())
+            .def(py::init([](const diskann::Metric metric, const size_t dim, const size_t max_points,
+                             const uint32_t complexity, const uint32_t graph_degree, const bool saturate_graph,
+                             const uint32_t max_occlusion_size, const float alpha, const uint32_t num_threads,
+                             const uint32_t filter_complexity, const uint32_t num_frozen_points,
+                             const uint32_t initial_search_complexity, const uint32_t search_threads,
+                             const bool concurrent_consolidation, const std::string &index_path) {
+                     return std::unique_ptr<DynamicInMemIndex<T>>(new DynamicInMemIndex<T>(
+                         metric, dim, max_points, complexity, graph_degree, saturate_graph, max_occlusion_size, alpha,
+                         num_threads, filter_complexity, num_frozen_points, initial_search_complexity, search_threads,
+                         concurrent_consolidation, index_path));
+                 }),
+                 py::arg("metric"), py::arg("dim"), py::arg("max_points"), py::arg("complexity"), py::arg("graph_degree"),
+                 py::arg("saturate_graph") = diskann::defaults::SATURATE_GRAPH,
+                 py::arg("max_occlusion_size") = diskann::defaults::MAX_OCCLUSION_SIZE,
+                 py::arg("alpha") = diskann::defaults::ALPHA, py::arg("num_threads") = diskann::defaults::NUM_THREADS,
+                 py::arg("filter_complexity") = diskann::defaults::FILTER_LIST_SIZE,
+                 py::arg("num_frozen_points") = diskann::defaults::NUM_FROZEN_POINTS,
+                 py::arg("initial_search_complexity") = 0, py::arg("search_threads") = 0,
+                 py::arg("concurrent_consolidation") = true, py::arg("index_path") = "")
+            .def("search", &DynamicInMemIndex<T>::search, py::arg("query"), py::arg("knn"), py::arg("complexity"))
+            .def("batch_search", &DynamicInMemIndex<T>::batch_search, py::arg("queries"), py::arg("num_queries"),
+                 py::arg("knn"), py::arg("complexity"), py::arg("num_threads"))
+            .def("insert", &DynamicInMemIndex<T>::insert, py::arg("vector"), py::arg("id"))
+            .def("mark_deleted", &DynamicInMemIndex<T>::mark_deleted, py::arg("id"))
+            .def("consolidate_delete", &DynamicInMemIndex<T>::consolidate_delete);
+
+        const std::string disk_name = "Disk" + class_name + "Index";
+        py::class_<DiskIndex<T>>(m, disk_name.c_str())
+            .def(py::init([](const diskann::Metric metric) {
+                return std::unique_ptr<DiskIndex<T>>(new DiskIndex<T>(metric));
+            }), py::arg("metric"))
+            .def("cache_bfs_levels", &DiskIndex<T>::cache_bfs_levels, py::arg("num_nodes_to_cache"))
+            .def("load_index", &DiskIndex<T>::load_index, py::arg("index_path_prefix"), py::arg("num_threads"),
+                 py::arg("num_nodes_to_cache"), py::arg("cache_mechanism") = 1)
+            .def("search", &DiskIndex<T>::search, py::arg("query"), py::arg("knn"), py::arg("complexity"),
+                 py::arg("beam_width"))
+            .def("batch_search", &DiskIndex<T>::batch_search, py::arg("queries"), py::arg("num_queries"),
+                 py::arg("knn"), py::arg("complexity"), py::arg("beam_width"), py::arg("num_threads"));
+}
+
 PYBIND11_MODULE(_diskannpy, m)
 {
     m.doc() = "DiskANN Python Bindings";
@@ -333,218 +432,28 @@ PYBIND11_MODULE(_diskannpy, m)
     m.attr("__version__") = "dev";
 #endif
 
-    m.def("build_in_memory_float_index", &build_in_memory_index<float>, py::arg("metric"), py::arg("data_file_path"),
-          py::arg("index_output_path"), py::arg("graph_degree"), py::arg("complexity"), py::arg("alpha"),
-          py::arg("num_threads"), py::arg("use_pq_build"), py::arg("num_pq_bytes"), py::arg("use_opq"),
-          py::arg("label_file") = "", py::arg("universal_label") = "", py::arg("filter_complexity") = 0);
-    m.def("build_in_memory_int8_index", &build_in_memory_index<int8_t>, py::arg("metric"), py::arg("data_file_path"),
-          py::arg("index_output_path"), py::arg("graph_degree"), py::arg("complexity"), py::arg("alpha"),
-          py::arg("num_threads"), py::arg("use_pq_build"), py::arg("num_pq_bytes"), py::arg("use_opq"),
-          py::arg("label_file") = "", py::arg("universal_label") = "", py::arg("filter_complexity") = 0);
-    m.def("build_in_memory_uint8_index", &build_in_memory_index<uint8_t>, py::arg("metric"), py::arg("data_file_path"),
-          py::arg("index_output_path"), py::arg("graph_degree"), py::arg("complexity"), py::arg("alpha"),
-          py::arg("num_threads"), py::arg("use_pq_build"), py::arg("num_pq_bytes"), py::arg("use_opq"),
-          py::arg("label_file") = "", py::arg("universal_label") = "", py::arg("filter_complexity") = 0);
+    // let's re-export our defaults
+    py::module_ default_values = m.def_submodule(
+        "defaults",
+        "A collection of the default values used for common diskann operations. `GRAPH_DEGREE` and `COMPLEXITY` are not"
+        " set as defaults, but some semi-reasonable default values are selected for your convenience. We urge you to "
+        "investigate their meaning and adjust them for your use cases.");
+
+    default_values.attr("ALPHA") = diskann::defaults::ALPHA;
+    default_values.attr("NUM_THREADS") = diskann::defaults::NUM_THREADS;
+    default_values.attr("MAX_OCCLUSION_SIZE") = diskann::defaults::MAX_OCCLUSION_SIZE;
+    default_values.attr("FILTER_COMPLEXITY") = diskann::defaults::FILTER_LIST_SIZE;
+    default_values.attr("NUM_FROZEN_POINTS") = diskann::defaults::NUM_FROZEN_POINTS;
+    default_values.attr("SATURATE_GRAPH") = diskann::defaults::SATURATE_GRAPH;
+    default_values.attr("GRAPH_DEGREE") = diskann::defaults::MAX_DEGREE;
+    default_values.attr("COMPLEXITY") = diskann::defaults::BUILD_LIST_SIZE;
+
+    add_variant<float>(m, "float", "Float");
+    add_variant<uint8_t>(m, "uint8", "UInt8");
+    add_variant<int8_t>(m, "int8", "Int8");
 
     py::enum_<Metric>(m, "Metric")
         .value("L2", Metric::L2)
         .value("INNER_PRODUCT", Metric::INNER_PRODUCT)
         .export_values();
-
-    py::class_<StaticInMemIndex<float>>(m, "DiskANNStaticInMemFloatIndex")
-        .def(py::init([](diskann::Metric metric,
-                         const std::string &data_path, // File location of vectors to be indexed
-                         const uint32_t l_build = diskann::defaults::BUILD_LIST_SIZE, // L build parameter
-                         const uint32_t build_max_degree = diskann::defaults::MAX_DEGREE) {
-            return std::unique_ptr<StaticInMemIndex<float>>(
-                new StaticInMemIndex<float>(metric, data_path, l_build, build_max_degree));
-        }))
-        .def("search", &StaticInMemIndex<float>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"))
-        .def("batch_search", &StaticInMemIndex<float>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("num_threads"));
-
-    py::class_<StaticInMemIndex<int8_t>>(m, "DiskANNStaticInMemInt8Index")
-        .def(py::init([](diskann::Metric metric,
-                         const std::string &data_path, // File location of vectors to be indexed
-                         const uint32_t l_build = diskann::defaults::BUILD_LIST_SIZE, // L build parameter
-                         const uint32_t build_max_degree = diskann::defaults::MAX_DEGREE) {
-            return std::unique_ptr<StaticInMemIndex<int8_t>>(
-                new StaticInMemIndex<int8_t>(metric, data_path, l_build, build_max_degree));
-        }))
-        .def("search", &StaticInMemIndex<int8_t>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"))
-        .def("batch_search", &StaticInMemIndex<int8_t>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("num_threads"));
-
-    py::class_<StaticInMemIndex<uint8_t>>(m, "DiskANNStaticInMemUInt8Index")
-        .def(py::init([](diskann::Metric metric,
-                         const std::string &data_path, // File location of vectors to be indexed
-                         const uint32_t l_build = diskann::defaults::BUILD_LIST_SIZE, // L build parameter
-                         const uint32_t build_max_degree = diskann::defaults::MAX_DEGREE) {
-            return std::unique_ptr<StaticInMemIndex<uint8_t>>(
-                new StaticInMemIndex<uint8_t>(metric, data_path, l_build, build_max_degree));
-        }))
-        .def("search", &StaticInMemIndex<uint8_t>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"))
-        .def("batch_search", &StaticInMemIndex<uint8_t>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("num_threads"));
-
-    py::class_<DynamicInMemIndex<float>>(m, "DiskANNDynamicInMemFloatIndex")
-        .def(py::init([](diskann::Metric metric, const size_t dim, const size_t max_points,
-                         const uint32_t l_build,          // L
-                         const uint32_t build_max_degree, // R
-                         const bool saturate_graph = diskann::defaults::SATURATE_GRAPH,
-                         const uint32_t max_occlusion_size = diskann::defaults::MAX_OCCLUSION_SIZE, // C
-                         const float alpha = diskann::defaults::ALPHA,
-                         const uint32_t num_threads = diskann::defaults::NUM_THREADS,
-                         const uint32_t filter_list_size = diskann::defaults::FILTER_LIST_SIZE, // Lf
-                         const uint32_t num_frozen_points = diskann::defaults::NUM_FROZEN_POINTS,
-                         const uint32_t initial_search_list_size = 0, const uint32_t search_threads = 1,
-                         const bool concurrent_consolidate = true) {
-            return std::unique_ptr<DynamicInMemIndex<float>>(new DynamicInMemIndex<float>(
-                metric, dim, max_points, l_build, build_max_degree, saturate_graph, max_occlusion_size, alpha,
-                num_threads, filter_list_size, num_frozen_points,
-                initial_search_list_size == 0 ? l_build : initial_search_list_size,
-                search_threads == 0 ? omp_get_num_threads() : search_threads,
-                concurrent_consolidate));
-        }))
-        .def("search", &DynamicInMemIndex<float>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"))
-        .def("batch_search", &DynamicInMemIndex<float>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("num_threads"))
-        .def("insert", &DynamicInMemIndex<float>::insert, py::arg("vector"), py::arg("id"))
-        .def("mark_deleted", &DynamicInMemIndex<float>::mark_deleted, py::arg("id"))
-        .def("consolidate_delete", &DynamicInMemIndex<float>::consolidate_delete);
-
-    py::class_<DynamicInMemIndex<int8_t>>(m, "DiskANNDynamicInMemInt8Index")
-        .def(py::init([](diskann::Metric metric, const size_t dim, const size_t max_points,
-                         const uint32_t l_build,          // L
-                         const uint32_t build_max_degree, // R
-                         const bool saturate_graph = diskann::defaults::SATURATE_GRAPH,
-                         const uint32_t max_occlusion_size = diskann::defaults::MAX_OCCLUSION_SIZE, // C
-                         const float alpha = diskann::defaults::ALPHA,
-                         const uint32_t num_threads = diskann::defaults::NUM_THREADS,
-                         const uint32_t filter_list_size = diskann::defaults::FILTER_LIST_SIZE, // Lf
-                         const uint32_t num_frozen_points = diskann::defaults::NUM_FROZEN_POINTS,
-                         const uint32_t initial_search_list_size = 0, const uint32_t search_threads = 1,
-                         const bool concurrent_consolidate = true) {
-            return std::unique_ptr<DynamicInMemIndex<int8_t>>(new DynamicInMemIndex<int8_t>(
-                metric, dim, max_points, l_build, build_max_degree, saturate_graph, max_occlusion_size, alpha,
-                num_threads, filter_list_size, num_frozen_points, initial_search_list_size == 0 ? l_build : initial_search_list_size,
-                search_threads == 0 ? omp_get_num_threads() : search_threads,
-                concurrent_consolidate));
-        }))
-        .def("search", &DynamicInMemIndex<int8_t>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"))
-        .def("batch_search", &DynamicInMemIndex<int8_t>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("num_threads"))
-        .def("insert", &DynamicInMemIndex<int8_t>::insert, py::arg("vector"), py::arg("id"))
-        .def("mark_deleted", &DynamicInMemIndex<int8_t>::mark_deleted, py::arg("id"))
-        .def("consolidate_delete", &DynamicInMemIndex<int8_t>::consolidate_delete);
-
-    py::class_<DynamicInMemIndex<uint8_t>>(m, "DiskANNDynamicInMemUInt8Index")
-        .def(py::init([](diskann::Metric metric, const size_t dim, const size_t max_points,
-                         const uint32_t l_build,          // L
-                         const uint32_t build_max_degree, // R
-                         const bool saturate_graph = diskann::defaults::SATURATE_GRAPH,
-                         const uint32_t max_occlusion_size = diskann::defaults::MAX_OCCLUSION_SIZE, // C
-                         const float alpha = diskann::defaults::ALPHA,
-                         const uint32_t num_threads = diskann::defaults::NUM_THREADS,
-                         const uint32_t filter_list_size = diskann::defaults::FILTER_LIST_SIZE, // Lf
-                         const uint32_t num_frozen_points = diskann::defaults::NUM_FROZEN_POINTS,
-                         const uint32_t initial_search_list_size = 0, const uint32_t search_threads = 1,
-                         const bool concurrent_consolidate = true) {
-            return std::unique_ptr<DynamicInMemIndex<uint8_t>>(new DynamicInMemIndex<uint8_t>(
-                metric, dim, max_points, l_build, build_max_degree, saturate_graph, max_occlusion_size, alpha,
-                num_threads, filter_list_size, num_frozen_points,
-                initial_search_list_size == 0 ? l_build : initial_search_list_size,
-                search_threads == 0 ? omp_get_num_threads() : search_threads,
-                concurrent_consolidate));
-        }))
-        .def("search", &DynamicInMemIndex<uint8_t>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"))
-        .def("batch_search", &DynamicInMemIndex<uint8_t>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("num_threads"))
-        .def("insert", &DynamicInMemIndex<uint8_t>::insert, py::arg("vector"), py::arg("id"))
-        .def("mark_deleted", &DynamicInMemIndex<uint8_t>::mark_deleted, py::arg("id"))
-        .def("consolidate_delete", &DynamicInMemIndex<uint8_t>::consolidate_delete);
-
-    py::class_<DiskANNIndex<float>>(m, "DiskANNFloatIndex")
-        .def(py::init([](diskann::Metric metric) {
-            return std::unique_ptr<DiskANNIndex<float>>(new DiskANNIndex<float>(metric));
-        }))
-        .def("cache_bfs_levels", &DiskANNIndex<float>::cache_bfs_levels, py::arg("num_nodes_to_cache"))
-        .def("load_index", &DiskANNIndex<float>::load_index, py::arg("index_path_prefix"), py::arg("num_threads"),
-             py::arg("num_nodes_to_cache"), py::arg("cache_mechanism") = 1)
-        .def("search", &DiskANNIndex<float>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"),
-             py::arg("beam_width"))
-        .def("batch_search", &DiskANNIndex<float>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("beam_width"), py::arg("num_threads"))
-        .def(
-            "build",
-            [](DiskANNIndex<float> &self, const char *data_file_path, const char *index_prefix_path, unsigned R,
-               unsigned L, double final_index_ram_limit, double indexing_ram_budget, unsigned num_threads,
-               unsigned pq_disk_bytes) {
-                std::string params = std::to_string(R) + " " + std::to_string(L) + " " +
-                                     std::to_string(final_index_ram_limit) + " " + std::to_string(indexing_ram_budget) +
-                                     " " + std::to_string(num_threads);
-                if (pq_disk_bytes > 0)
-                {
-                    params = params + " " + std::to_string(pq_disk_bytes);
-                }
-                diskann::build_disk_index<float>(data_file_path, index_prefix_path, params.c_str(), self.get_metric());
-            },
-            py::arg("data_file_path"), py::arg("index_prefix_path"), py::arg("R"), py::arg("L"),
-            py::arg("final_index_ram_limit"), py::arg("indexing_ram_limit"), py::arg("num_threads"),
-            py::arg("pq_disk_bytes") = 0);
-
-    py::class_<DiskANNIndex<int8_t>>(m, "DiskANNInt8Index")
-        .def(py::init([](diskann::Metric metric) {
-            return std::unique_ptr<DiskANNIndex<int8_t>>(new DiskANNIndex<int8_t>(metric));
-        }))
-        .def("cache_bfs_levels", &DiskANNIndex<int8_t>::cache_bfs_levels, py::arg("num_nodes_to_cache"))
-        .def("load_index", &DiskANNIndex<int8_t>::load_index, py::arg("index_path_prefix"), py::arg("num_threads"),
-             py::arg("num_nodes_to_cache"), py::arg("cache_mechanism") = 1)
-        .def("search", &DiskANNIndex<int8_t>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"),
-             py::arg("beam_width"))
-        .def("batch_search", &DiskANNIndex<int8_t>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("beam_width"), py::arg("num_threads"))
-        .def(
-            "build",
-            [](DiskANNIndex<int8_t> &self, const char *data_file_path, const char *index_prefix_path, unsigned R,
-               unsigned L, double final_index_ram_limit, double indexing_ram_budget, unsigned num_threads,
-               unsigned pq_disk_bytes) {
-                std::string params = std::to_string(R) + " " + std::to_string(L) + " " +
-                                     std::to_string(final_index_ram_limit) + " " + std::to_string(indexing_ram_budget) +
-                                     " " + std::to_string(num_threads);
-                if (pq_disk_bytes > 0)
-                    params = params + " " + std::to_string(pq_disk_bytes);
-                diskann::build_disk_index<int8_t>(data_file_path, index_prefix_path, params.c_str(), self.get_metric());
-            },
-            py::arg("data_file_path"), py::arg("index_prefix_path"), py::arg("R"), py::arg("L"),
-            py::arg("final_index_ram_limit"), py::arg("indexing_ram_limit"), py::arg("num_threads"),
-            py::arg("pq_disk_bytes") = 0);
-
-    py::class_<DiskANNIndex<uint8_t>>(m, "DiskANNUInt8Index")
-        .def(py::init([](diskann::Metric metric) {
-            return std::unique_ptr<DiskANNIndex<uint8_t>>(new DiskANNIndex<uint8_t>(metric));
-        }))
-        .def("cache_bfs_levels", &DiskANNIndex<uint8_t>::cache_bfs_levels, py::arg("num_nodes_to_cache"))
-        .def("load_index", &DiskANNIndex<uint8_t>::load_index, py::arg("index_path_prefix"), py::arg("num_threads"),
-             py::arg("num_nodes_to_cache"), py::arg("cache_mechanism") = 1)
-        .def("search", &DiskANNIndex<uint8_t>::search, py::arg("query"), py::arg("knn"), py::arg("l_search"),
-             py::arg("beam_width"))
-        .def("batch_search", &DiskANNIndex<uint8_t>::batch_search, py::arg("queries"), py::arg("num_queries"),
-             py::arg("knn"), py::arg("l_search"), py::arg("beam_width"), py::arg("num_threads"))
-        .def(
-            "build",
-            [](DiskANNIndex<uint8_t> &self, const char *data_file_path, const char *index_prefix_path, unsigned R,
-               unsigned L, double final_index_ram_limit, double indexing_ram_budget, unsigned num_threads,
-               unsigned pq_disk_bytes) {
-                std::string params = std::to_string(R) + " " + std::to_string(L) + " " +
-                                     std::to_string(final_index_ram_limit) + " " + std::to_string(indexing_ram_budget) +
-                                     " " + std::to_string(num_threads);
-                if (pq_disk_bytes > 0)
-                    params = params + " " + std::to_string(pq_disk_bytes);
-                diskann::build_disk_index<uint8_t>(data_file_path, index_prefix_path, params.c_str(),
-                                                   self.get_metric());
-            },
-            py::arg("data_file_path"), py::arg("index_prefix_path"), py::arg("R"), py::arg("L"),
-            py::arg("final_index_ram_limit"), py::arg("indexing_ram_limit"), py::arg("num_threads"),
-            py::arg("pq_disk_bytes") = 0);
 }
